factor script optimization Expr AST nodes out of main AST header
