+++ src/Expr.cc
@@ -24,7 +24,7 @@
 #include "zeek/broker/Data.h"
 #include "zeek/digest.h"
 #include "zeek/module_util.h"
-#include "zeek/script_opt/ExprOptInfo.h"
+#include "zeek/script_opt/Expr.h"
 #include "zeek/script_opt/ScriptOpt.h"
 
 namespace zeek::detail {
