+++ src/script_opt/CPP/Exprs.cc
@@ -481,7 +481,8 @@ string CPPCompile::GenAddToExpr(const Expr* e, GenType gt, bool top_level) {
     if ( t->Tag() == TYPE_VECTOR ) {
         auto& rt = rhs->GetType();
 
-        if ( e->AsAddToExpr()->IsVectorElemAppend() )
+        ASSERT(e->Tag() == EXPR_ADD_TO);
+        if ( static_cast<const AddToExpr*>(e)->IsVectorElemAppend() )
             add_to_func = "vector_append__CPP";
         else
             add_to_func = "vector_vec_append__CPP";
